# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HostOnboardingRequest": {
        "properties": {
          "payout_iban": {
            "maxLength": 34,
            "minLength": 15,
            "title": "Payout Iban",
            "type": "string"
          }
        },
        "required": [
          "payout_iban"
        ],
        "title": "HostOnboardingRequest",
        "type": "object"
      },
      "HostProfileOut": {
        "properties": {
          "onboarding_state": {
            "title": "Onboarding State",
            "type": "string"
          },
          "payout_iban_masked": {
            "title": "Payout Iban Masked",
            "type": "string"
          },
          "stripe_account_id": {
            "title": "Stripe Account Id",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "onboarding_state",
          "stripe_account_id",
          "payout_iban_masked"
        ],
        "title": "HostProfileOut",
        "type": "object"
      },
...
      "LegalIdentityIn": {
        "description": "The legal name ownership is checked against — as it appears in the land\nregister, not a display name.",
        "properties": {
          "country_of_residence": {
            "anyOf": [
              {
                "pattern": "^[A-Z]{2}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Of Residence"
          },
          "legal_first_name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Legal First Name",
            "type": "string"
          },
          "legal_last_name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Legal Last Name",
            "type": "string"
          }
        },
        "required": [
          "legal_first_name",
          "legal_last_name"
        ],
        "title": "LegalIdentityIn",
        "type": "object"
      },
      "LegalIdentityOut": {
        "properties": {
          "country_of_residence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Of Residence"
          },
          "legal_first_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legal First Name"
          },
          "legal_last_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legal Last Name"
          },
          "legal_party_id": {
            "title": "Legal Party Id",
            "type": "string"
          }
        },
        "required": [
          "legal_party_id",
          "legal_first_name",
          "legal_last_name",
          "country_of_residence"
        ],
        "title": "LegalIdentityOut",
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "PhoneVerificationStart": {
        "description": "E.164 only. A free-text number cannot be compared for uniqueness, and\nwithout uniqueness \"verified phone\" stops being one account per SIM.",
        "properties": {
          "phone": {
            "maxLength": 16,
            "minLength": 8,
            "pattern": "^\\+[1-9]\\d{7,14}$",
            "title": "Phone",
            "type": "string"
          }
        },
        "required": [
          "phone"
        ],
        "title": "PhoneVerificationStart",
        "type": "object"
      },
...
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest",
        "type": "object"
      },
      "RegisterRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "full_name": {
            "default": "",
            "title": "Full Name",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 10,
            "title": "Password",
            "type": "string"
          },
          "role": {
            "default": "guest",
            "enum": [
              "guest",
              "host"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
...
      "TokenPair": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "expires_in"
        ],
        "title": "TokenPair",
        "type": "object"
      },
      "UserOut": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "email_verified_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email Verified At"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "phone_verified_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone Verified At"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "full_name",
          "role"
        ],
        "title": "UserOut",
        "type": "object"
      },
...
      "VerificationStarted": {
        "properties": {
          "channel": {
            "enum": [
              "email",
              "phone"
            ],
            "title": "Channel",
            "type": "string"
          },
          "destination_masked": {
            "title": "Destination Masked",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          }
        },
        "required": [
          "channel",
          "destination_masked",
          "expires_in"
        ],
        "title": "VerificationStarted",
        "type": "object"
      },
      "VerificationState": {
        "properties": {
          "email_verified": {
            "title": "Email Verified",
            "type": "boolean"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "phone_verified": {
            "title": "Phone Verified",
            "type": "boolean"
          }
        },
        "required": [
          "email_verified",
          "phone_verified"
        ],
        "title": "VerificationState",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Managed hospitality platform. This spec is generated from the implementation (code-first) and kept in sync by a CI drift-guard — see docs/api/README.md.",
    "title": "Homies API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/auth/login": {
      "post": {
        "operationId": "login_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "operationId": "refresh_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/auth/register": {
      "post": {
        "operationId": "register_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register",
        "tags": [
          "identity"
        ]
      }
    },
...
    "/v1/hosts/me": {
      "get": {
        "operationId": "host_me_v1_hosts_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostProfileOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Host Me",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/hosts/onboarding": {
      "post": {
        "operationId": "host_onboarding_v1_hosts_onboarding_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostOnboardingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostProfileOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Host Onboarding",
        "tags": [
          "identity"
        ]
      }
    },
...
    "/v1/me": {
      "get": {
        "operationId": "me_v1_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Me",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/me/legal-identity": {
      "put": {
        "description": "Record the legal name a property claim is verified against.\n\nLocked once any claim held under it has been verified. Otherwise a\nverified owner could rename themselves afterwards, and the verification\nwould silently vouch for a person nobody checked.",
        "operationId": "set_legal_identity_v1_me_legal_identity_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LegalIdentityIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegalIdentityOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Legal Identity",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/me/notifications": {
      "get": {
        "description": "In-app notification feed for the logged-in user (guest or host).",
        "operationId": "my_notifications_v1_me_notifications_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "My Notifications",
        "tags": [
          "identity"
        ]
      }
    },
...
    "/v1/me/verification": {
      "get": {
        "operationId": "verification_state_v1_me_verification_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationState"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Verification State",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/me/verify/email/start": {
      "post": {
        "operationId": "start_email_verification_v1_me_verify_email_start_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationStarted"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Start Email Verification",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/me/verify/phone/start": {
      "post": {
        "operationId": "start_phone_verification_v1_me_verify_phone_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhoneVerificationStart"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationStarted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Start Phone Verification",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/me/verify/{channel}/confirm": {
      "post": {
        "operationId": "confirm_verification_v1_me_verify__channel__confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "channel",
            "required": true,
            "schema": {
              "title": "Channel",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerificationConfirm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Confirm Verification",
        "tags": [
          "identity"
        ]
      }
    },
...
  },
...
}